feat: se agregan validaciones

+++ app/Http/Controllers/CardTokenController.php
@@ -15,7 +15,8 @@ class CardTokenController extends Controller
         private CardTokenService $cardTokenService
     ) {}
 
-    public function indexClient() {
+    public function indexClient()
+    {
         $client = auth()->user()->client;
         if (!$client) return ApiResponseClass::sendResponse(null, "Cliente encontrado", 404);
 
@@ -24,7 +25,8 @@ class CardTokenController extends Controller
         return ApiResponseClass::sendResponse(CardTokenResource::collection($data));
     }
 
-    public function tokenizeCard(CreateCardTokenRequest $request) {
+    public function tokenizeCard(CreateCardTokenRequest $request)
+    {
         $client = auth()->user()->client;
         if (!$client) return ApiResponseClass::sendResponse(null, "Cliente encontrado", 404);
 
@@ -46,23 +48,52 @@ class CardTokenController extends Controller
         }
     }
 
-    public function updateTokenizedCard(UpdateCardTokenRequest $request, $id) 
+    public function updateTokenizedCard(UpdateCardTokenRequest $request, $id)
     {
+        $client = auth()->user()->client;
+        if (!$client) return ApiResponseClass::sendResponse(null, "Cliente encontrado", 404);
+
         $data = $request->only([
             'card',
         ]);
 
-        $cardToken = $this->cardTokenService->updateTokenizedCard($id, $data);
-        if (!$cardToken) return ApiResponseClass::sendResponse(null, "Token de tarjeta con ID $id no encontrado", 404);
+        DB::beginTransaction();
 
-        return ApiResponseClass::sendResponse(new CardTokenResource($cardToken));
+        try {
+            $cardToken = $this->cardTokenService->updateTokenizedCard($id, $data);
+            if (!$cardToken) return ApiResponseClass::sendResponse(null, "Token de tarjeta con ID $id no encontrado", 404);
+
+            if ($cardToken->client_id !== $client->id) {
+                return ApiResponseClass::sendResponse(null, "Tarjeta no valida", 403);
+            }
+
+            DB::commit();
+            return ApiResponseClass::sendResponse(new CardTokenResource($cardToken));
+        } catch (\Exception $ex) {
+            return ApiResponseClass::rollback($ex);
+        }
     }
 
-    public function deleteTokenizedCard($id) 
+    public function deleteTokenizedCard($id)
     {
-        $cardToken = $this->cardTokenService->deleteTokenizedCard($id);
-        if (!$cardToken) return ApiResponseClass::sendResponse(null, "Token de tarjeta con ID $id no encontrado", 404);
+        $client = auth()->user()->client;
+        if (!$client) return ApiResponseClass::sendResponse(null, "Cliente encontrado", 404);
 
-        return ApiResponseClass::sendResponse(new CardTokenResource($cardToken));
+        DB::beginTransaction();
+
+        try {
+            $cardToken = $this->cardTokenService->deleteTokenizedCard($id);
+            if (!$cardToken) return ApiResponseClass::sendResponse(null, "Token de tarjeta con ID $id no encontrado", 404);
+
+
+            if ($cardToken->client_id !== $client->id) {
+                return ApiResponseClass::sendResponse(null, "Tarjeta no valida", 403);
+            }
+
+            DB::commit();
+            return ApiResponseClass::sendResponse(new CardTokenResource($cardToken));
+        } catch (\Exception $ex) {
+            return ApiResponseClass::rollback($ex);
+        }
     }
 }
